successful integration of authentication with the frontend
