--- src/modules/authentication/individualUserAuth/googleIndividualUser/googleIndividualUserAuth.route.ts
@@ -1,38 +1,0 @@
-import express from "express";
-import {
-  // authWithGoogle,
-  getGoogleUrl,
-  getGoogleUserDetail,
-} from "./googleIndividualUserAuth.controller";
-
-const router = express.Router();
-/**
- * @swagger
- * tags:
- *  name: GoogleIndividualUserAuth
- * description: To signup and login google individual user
- */
-
-/**
- * @swagger
- *   /auth/individual/oauth:
- *     post:
- *       summary: Get google authorized url; navigate to google consent page; and give google user access
- *       tags: [GoogleIndividualUserAuth]
- *       responses:
- *         '200':
- *           description: Google AuthorizedUrl successful gotten
- *           content:
- *             application/json:
- *               schema:
- *                 $ref: "#/components/schemas/GoogleOrganizationAccess"
- *         '401':
- *           $ref: "#/components/responses/401"
- */
-
-// router.post("/oauth", getGoogleUrl);
-router.get("/oauth", getGoogleUrl);
-router.post("/oauth/callback", getGoogleUserDetail);
-// router.post("/oauth/callback", authWithGoogle);
-
-export default router;
