+++ src/modules/authentication/individualUserAuth/individualAuthPasswordToken.ts
